py-interface/requirements.txt
numpy==2.4.3
rasterio==1.5.0
geopandas==1.1.3
sentinelhub==3.11.4
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.2.2
pyproj==3.7.2
PyYAML==6.0.3
